--- src/MataAtlantica.API/Middleware/ExceptionMiddleware.cs
@@ -1,59 +1,0 @@
-using FluentValidation;
-using System.Text.Json;
-
-namespace MataAtlantica.API.Middleware;
-
-public class ExceptionMiddleware : IMiddleware
-{
-    private readonly ILogger<ExceptionMiddleware> _logger;
-    public ExceptionMiddleware(ILogger<ExceptionMiddleware> logger) => _logger = logger;
-    public async Task InvokeAsync(HttpContext context, RequestDelegate next)
-    {
-        try
-        {
-            await next(context);
-        }
-        catch (Exception e)
-        {
-            _logger.LogError(e, e.Message);
-            await HandleExceptionAsync(context, e);
-        }
-    }
-    private static async Task HandleExceptionAsync(HttpContext httpContext, Exception exception)
-    {
-        var statusCode = GetStatusCode(exception);
-        var response = new 
-        {
-            title = GetTitle(exception),
-            status = statusCode,
-            detail = exception.Message,
-            errors = GetErrors(exception)
-        };
-        httpContext.Response.ContentType = "application/json";
-        httpContext.Response.StatusCode = statusCode;
-        await httpContext.Response.WriteAsync(JsonSerializer.Serialize(response));
-    }
-    private static int GetStatusCode(Exception exception) =>
-        exception switch
-        {
-            ValidationException => StatusCodes.Status400BadRequest,
-            _ => StatusCodes.Status500InternalServerError
-        };
-    private static string GetTitle(Exception exception) =>
-        exception switch
-        {
-            // ApplicationException applicationException => applicationException.Title,
-            _ => "Server Error"
-        };
-    private static IReadOnlyDictionary<string, string> GetErrors(Exception exception)
-    {
-        Dictionary<string, string> errors = null;
-        if (exception is ValidationException validationException)
-        {
-            errors = validationException
-                .Errors
-                .ToDictionary(p => p.ErrorCode, p => p.ErrorMessage);
-        }
-        return errors;
-    }
-}
